Reduce memory needed by IconColorEditor
IconColorEditor was building all icons with a size of 512x512 pixels
which takes a lot of memory especially considering that in reality the icons are shown as 32x32 pixel images. This commit splits a new class IconListManager from IconManager. IconListManager has the code needed only for IconColorEditor and it always draws the icons in requested size (32x32). What was left in IconManager works as before with Graph and Tree views.

Re #357