Mock locale functions in test_CopyPasteTableView
The tests in the file are supposed to run in an environment where the
decimal separator differs from '.'. The previous implementation where
the locale was set to 'DE_de' didn't work universally because that
specific locale may not be available on every system. We now mock some
specific functions in the locale module instead to work around the problem.

Re #361