Fix scrolling in Design view
CustomGraphicsView and DesignGraphicsView were using their own
setSceneRect() functions when new items were dragged and dropped on the
Design view. This function, however, does not affect the underlying
graphics scene but rather modifies the region shown by the Design view.
This resulted in the scroll bars not showing the entire scene as the
scene itsef had a different rect.

We now leave the graphics view's scene rect untouched (i.e. leave
it as None) which makes Qt to automatically update the region shown in
Desing view according to changes in the scene.

Re #362