Improve zooming in Desing view
Design view had a regression which caused zooming by mouse wheel to work
in an unintuitive way. This was because the scene view was centered on
the mouse after scaling which caused the view to seemingly jump around
wildly. We now center the view post-scaling such that the point in the
scene below the mouse cursor always stays in the same place. This should
provide a smoother zooming experience.

Re #362