Fix zooming when initial scale out of bounds
There are situations where the Graph view may scale a graph such that
the scaling factor is out of the nominal scaling limits of
CustomQGraphicsView to fit everything in the view. In these situations
it should still be possible to zoom in (or out) to get to the
'acceptable' scale range.

Re #362