Disable zoom in Graph view while dragging an item
The graphics view was automatically zooming out when an item was dragged
outside the visible area. This was in principle a nice feature but led
to the item moving uncontrollably farther away from its original
position. Zooming in such a situation is now disabled. The background
item in the Graph view has been removed. We update the scene rect
instead.

Re #362