Fix Desing view zoom at startup
When a previous project was opened automatically at Toolbox startup and
the item scene was small enough (e.g. a single item), zoom in the view
was way off. This was because the call to reset_zoom() happened before
Design view had everything set up correctly. We now call init_project()
after calling show() for the main window instead of calling it in
__init__(). At that point Desing view dimensions are set up correctly
and the initial zoom level can be calculated correctly.

Re #362