Move the `project_items` pkg from `plugins` to `spinetoolbox`
Our official project items don't need to be shipped as plugins. They are 
better as an integral part of the `spinetoolbox` pkg. That said, it 
should still be possible to load a project item pkg from any directory 
as a plugin, but that will be done for community-developed project 
items.