Introduce ProjectItemResource has multipurpose resource holder
We want to be flexible on what type of data project items can advertise 
as resource. So instead of holding one dictionary in the 
ExecutionInstance for each currently supported resource, we hold a 
unique dictionary for all resources which need to be instances of 
ProjectItemResource. ProjectItemResource has an attribute called `type_` 
that is supposed to say what this resource is about. So ProjectItems can 
check the `type_` of the available resources and react accordingly. 
Future project items can follow the same pattern which provides ultimate 
flexibility as `type_` is just a string.

Re issue #398