Fix method for creating a new Spine db
This got incrementally broken as we introduced the notification icon 
thing (see issue #343). One of the requirements was to show a 
notification icon on Data Store items whenever the url was incorrect. So 
everytime the user changes the selection in the Data Store properties, 
we update the `self._url` attribute and check whether or not the new url 
is valid. This is fine.

Now, the method that creates the new spine db first tries to make a url 
from the current selections and if it fails, it *sets* the selections to 
the defaults and tries again. But *setting* the selections triggers the 
other process that checks the url for the notification icon thing.

The fix is to set the selections in a way that doesn't trigger the other 
process about the notification icon. Better to keep this separate, for 
clarity.

Tests have ben updated in a previous commit (49fa8125) but more will be 
added in a future commit.

Re issue #379