Add rudimentary exception handling
GamsExceptions are now confided within the gdx module and converted to
GdxExportException classes --- we cannot let other parts of Spinetoolbox
to try importing gams and having to deal with the problems in case the
Python bindings are not found. Gdx Export item also now catches these
exceptions and aborts execution gracefully.

Re #274