Update url attribute in data store when db changes
We were relying on QLineEdit.setText to automatically update the url 
attribute in Data Stores. However, we recently changed the behaviour so 
`setText` doesn't do that anymore. The reason is we no longer connect 
`textChanged` to the method that updates the url attribute, because that 
means the attrbute is updated everytime the user types a new letter in 
the text edit (and this triggers the project simulation).

So simply, we just update the url attribute manually whenever the user 
selects a new file.

Should solve #412