Improve `ProjectItemResource`
Instead of hard-coding project items into the `type_` attribute, e.g., 
"tool_output_file", we now just say "file". On top of that, we add the 
`url` attribute used for both "database" *and* "file" types.

Moreover, we add the `metadata` attribute to convey extra information 
about the resource. Project items can query this metadata to know 
exactly if and how to interact with the resouce.

Re Issue #398