Refactor file copying in Tool and ToolInstance
The Tool class was taking care of copying input files to the work (or 
source) directory, whereas the ToolInstance class was taking care of 
copying source files to the work directory (if needed) and output files 
to the result directory. This distribution of work seemed arbitrary, so 
now the Tool class takes care of all file-copying (input, optional 
input, source, and output files).

Output file resource advertising is also done in the Tool class now, 
which also seems like a better place for that.

A new test class is provided, for testing the execution of two arbitrary 
tool templates, one simple and one complex. The complex includes 
wildcards in the optional input files and soruce files so we test some 
pattern matching too.

Re issue #417