Introduce ToolInstance subclasses
It seemed more pythonic to have different ToolInstance subclasses with 
an specialized `execute` method, rather than a 'switch' on the tool type 
within the `execute` method of `ToolInstance`. The above also applied to 
the method `update_instance` of the Tool class. So now all this have 
been refactored the more pythonic way. It looks slightly better.

(Note the method `create_tool_instance` added to `ToolTemplate` to 
obtain an instance of the right `ToolInstance` subclass.)

Re issue #417