Fix Gdx Export settings window closing
There were situations in which the window was removed from GdxExport's
open window list before the modified settings were stored back to the
item. This was due to the OK button being connected to close() and
_approve_settings() slot at the same time; the order in which these were
invoked was not always clear. The button is now connected to GdxExport
item only which calls close() after the modified settings have been
processed.

Re #274