Log execution start before errors
Gdx Export item was bailing out with an error message immediately when
an issue with GAMS Python bindngs was detected making it look like the
previous item in a DAG was having issues. We now print 'Executing Gdx
Export...' message before checking for errors so users know which item
is actually failing.

Re #274