Replace RuntimeException by GdxExportException
Raise GdxExportException instead of the too-general RuntimeException and
do this in gdx.make_gams_workspace() instead of to_gams_workspace().
Also, except this exception in unit tests too and skip such tests to
allow the tests to run on systems with broken GAMS installations (not an
error in Toolbox).

Re #274