Introduce ParameterItem into the empty models
ParameterItem is a new class to populate rows in the parameter models. 
The empty parameter models have been refactored to use this class. This 
allows way more flexibility and isolation. Now everything that's common 
to all entities is done in a base class model, whereas specifities of 
each entity type are handled in subclasses of that base class.

Re issue #415