Refactor project items
- Move item_type to ProjectItem base class from the concere project
items. All of them should have this attribute and this is the way to
enforce it.
- Add notify_destination() method to ProjectItem. This moves the
responsibility of informing user of item connections from
DesignQGraphicsView to the items themselves leading to less coupling,
clear interfaces and less spaghetti code.
- Added unit tests for item_type and notify_destination(). This means
that we now have at least two tests for each concrete project item.

Re #413