Fix bugs and clean up
After the big refactoring in one past commit we needed to checkout if 
everything was working as expected, and of course it wasn't. So now we 
clean up and fix many bugs so basic functionality is there: at least 
it's possible to edit all parameter tables, as well as to add and update 
items. Filtering from the tress also work.

Support for autofiltering is pending, as well as removal, renaming from 
the tree, ect.

Re #415