Implemente new filter technique in parameter models
We weren't using enough from QSortFilterProxyModel in the Compound model 
to see real benefit from it. Plus there was some overhead from calling 
sourceModel() repeteadly. For this reason, I decided to reimplemented 
the basic row filter functionality on my own.

The result is exactly the same and hopefully it simplifies/optimizes the 
code.

Re #415