Fix bug in tool's simulate execution
When simulating the tool we were attempting to create a uri from a 
relative path which doesn't work. Now ProjectItemResource works with 
urls we need to kinda invent a full path for that 'simulated' file in 
order to work. It's a bit of overhead but nothing dramatic imo. 

Also import os in view.

Re #398