More informative dialog button labels
Changed some 'Yes' and 'No' buttons to 'Some Action' and 'Cancel'
buttons to give a quick user a change to actually read what they are
doing before hitting the first button in sight. Also, closing one of the
data store views not behave a bit more naturally with regards to
prompting if databases changes should be committed or not.

Re #293