Let items advertise data only to their direct neighbors
Until now, execution would propagate data so eg a file advertised by a 
DC at the beginning of the DAG would be seen by *all* the items down the 
line. 

This has its benefits but can also lead to really unexpected behavior so 
the decision is to stop doing it. Data is advertised one level further 
in the DAG and not beyond.

Should solve #405