Change the way we handle console execution failure
Affects both Python and Julia tools executed in console mode.
In case execution failed, we were relaunching the process this time
using the command prompt. This is now disabled as it makes little sense:
the reason the tool failed is probably not the fact it was
executed in the console... So now if the tool fails we inform the
user and that's that.

Re #417