Improve Graph view zooming
Scene rect would grow indefinitely when repeatedly zooming out to the
max zoom as the items were manually scaled to match the zoom level; this
was not always accurate and caused the items to actually scale slowly.
Now, most items have the ItemIgnoresTransformations flag set except
ArcItem for which we keep the line width fixed manually. The new
implementation seems to perform faster, too. A drawback is that moving
object items has to be done manually now as without the transformations they
don't always follow the mouse pointer properly.

Re #444