Fix scene resizing in Graph View
QGraphicsScene in Graph View failed to resize it's scene rect when
moving an item and:
- There was a connected arc with zero length. This could happen when the
source and destination ObjectItems were the same
- There was an ArcTokenItem which did not hold any object information
In both cases the bounding rect of the items was empty somehow resulting
in the scene not understanding that its bounds were crossed. Anyway, we
don't draw such arcs or arc tokens anymore since they really don't carry
any information to the user.

Re #452