Draw Arc tokens properly in graph view
We were drawing all dimensions in the arc tokens where actually we
only need to draw the dimensions other than the endpoints.

This also allows us to draw tokenless arcs, in which case we set
a minimal size QRect to the token so the shrinking scene can keep
keeping track.

Re #415 and #452