(WIP) Try to solve bug when closing and opening tree view forms
Turns out PySide2 wrappers around QObjects can outlive them, or
something like that.

The bug we see is the following: open a tree view, close it,
open another tree view or even a graph view and do whatever
action that triggers a SpineDBManager signal. The
Python interpreter complains, "Internal C++ object already deleted."

Apparently, the form that was closed is still alive somehow,
with all its slots still listening to the SpineDBManager signals.
The QObject is destroyed (we receive the `destroyed` signal)
but there are still plenty of Python references to it.
Its slots won't be disconnected until the object is garbage
collected.

I guess we could try and manually disconnect all signals from slots
in the close event, but that's horrible.
To me, the only acceptable solution is to make sure all references
are deleted when closing the form. This requires a lot of attention,
or at least all the attention I haven't paid so far when storing
references to an object.

I believe I'm close now, but there are still some corner cases
where the form doesn't die after closing it.

Re #415