Remove all trailing references to tree view form upon closing it
This was harder than I expected, but finally not without a few dirty
tricks I was able to delete all references to the tree and graph view
forms when they're closed. This means all slots are disconnected
and we don't see a traceback when making the Spine DB manager work
after a tree or graph view form has been closed.

Re #415