Fix pasting in parameter tables
Turns out we were largely behind here. First of all, the compound
table model didn't have a insertRows method, so the table couldn't
grow upon pasting. Now the method is added and seems to work fine.

Second, we were assigning the same list as default row to several rows
in the empty model, and we know how Python likes that. Now it's fixed.

And last, entering gibberish in the value list and parameter tag
columns was being silently ignored. Now we emit error messages so the
user knows is their fault and not ours.

Re #415