Introduce new class to handle spine db manager signals
Turns out that one shouldn't rely on object deletion for anything
in Python (see https://docs.python.org/2/reference/datamodel.html).
Basically when the reference count hit zero the object may be claimed
or *maybe not*.

So no choice for us, we need to manually handle the problem.
(Btw, I'm talking here about the traceback with the message
'Internal C++ object already deleted', due to closed tree view forms
keep receiving signals from the db manager)

So I've created a new class called SpineDBSignaller.
SpineToolboxProject now also has an attribute of this class.
Components that want to react to SpineDBManager signals have to
subscribe as listeners in this signaller, and then the signaller c
alls a predefined method in the component when the signal happens.
In the close event, the component needs to unsubscribe to stop
being part of the process. This should solve the issue more
consistently.

Re #415