Overwrite items in cache when updating
Re #463

With 5c774cf3, accessing a parameter value adds a new key
to the cached dictionary containing the parsed value. Following
accesses look at this key first to avoid multiple parsing.

So now, when the value is updated we need to delete that key
so the value is refreshed. This is now done by replacing the cache
entry (before we were just updating it).