Control db creation, commit and rollback in SpineDBManager
Re #465

Weird things happened when creating new spine dbs on urls
that were being monitored by the SpineDBManager. Now the manager is
in charge of creating the new spine database, so it can check if
the given url is one of the monitored ones, close the db map
connections and notify the tree view. For now, the tree view reacts
by prompting the user and gracefully closing.

In addition, the DB manager also controls commit and rollback actions
so this can be synched across widgets.