Draw relationships in graph view as nodes (WIP)
Re #455

Until, the relationship in the graph view was represented by means
of the arcs. Each arc represented a segment connecting two successive
members in the relationship. This had multiple drawbacks as it was
hard to represent eg a relationship between an object an itself,
or the same object appearing more than once in one relationship.

So now we draw relationships as nodes and it makes a whole lot more
sense. Arcs are drawn between the relationship node and each member
node. With this the user has a better view on the relationship and
overall the graph looks better.

This is still a decent-size refactoring, but this commit provides
the core. Some details will be finished in further commits.