Enable commit and rollback actions at tree view form load time
Before the big refactoring of #415, all actions started disabled and
we enabled them as the user did things.
This seemed to be too much maintenance so now we only
enable or disable actions in a Slot connected to the aboutToShow
signal of the corresponding menu.
But sometimes the user uses the shortcut to trigger the action, so
we also need to enable the actions at form load.