Emit the right signal when updating parameter definitions and values
This is again some detail that we missed when wrapping up #415.
Near the end of the refactoring we introduced an internal signal-slot
connection in SpineDBManager, so when parameter definitions and values
are added or updated we can go from compact to extend format.

The difference between compact and extended is the latter has
entity and class information, as it is displayed in the tree and graph
view parameter tables.