Fix bug when showing parameter tables
Sometimes it happened that one couldn't see any rows in the
parameter tables even if data was there. I believe this was due
to emitting rowsInserted too soon -- before updating the internal
data structure of the model. Hopefully it's solved now. I say
hopefully because the bug was kinda non-deterministic.