Close db_map connection when all views close
Re #470

Before this commit, when opening any of the Data Store views we
created the necessary db_map and kept it alive basically forever
(until the app exited). This seemed sloppy, and had some real drawbacks
like making difficult to delete the SQLite file on windows.

Now we close the db_map as soon as the last form viewing it is closed.
If there are pending changes, the 'commit or discard' prompt is
shown at that moment too. The implementation relies on the
SpineDBSignaller. Basically the trick is to keep track of which db_maps
are being listened to, and then if nobody is listening to a db_map
to react accordingly.

The tabular view needs more work, but that will probably be tackled
within issue #461. With this commit, we'll probably see tracebacks
when using the tabular view since we register it as a listener,
but haven't implemented any of the handlers for the DB manager signals.