Try to fix tree view tests
Travis is hanging forever on tree view tests, and I'm pretty sure
it's because a QMessageBox is being executed. Now I'm mocking
QMessageBox in the tearDown method so hopefully tests will pass.

In my machine tests pass without mocking QMessageBox but it's
probably because of my local QSettings. Let's see.