Advertise real resources in Tool simulation
In the `simulate_execution` method of `Tool`, we were creating
artificial file paths just by joining the output dir with every
file name in the specification's output file list. These never
correspond to actual output files which are actually in a
timestamped subfolder. As result, the view item for instance was
seeing a file that didn't exist, and execution was needed to correct
this problem.

Now, during simulation the Tool walks its output directory and
advertises every relevant file in there.