Fix superfluous error logging while adding items
Some project items trigger the item_changed signal durint __init__. This was
causing SpineToolboxProject.notify_all_items_of_dag_changes() to get
called before the item was actually in a DAG causing an error message in
the logs. We now print the message only if the item is in the project
item model before showing the error.

Re #418