Add ExecutionState.WAIT state
Previously, Tool._do_execute() did not return a ExecutionState value but
rather None indicating that the execution was still continuing (in another
process/thread). 'None' is not a very descriptive object for an
execution state so this commit introduces ExecutionState.WAIT to
indicate that a project item is still processing although its execute()
has finished.

Re #418