Use own DatabaseMapping in Exporter
Exporter is now using its own DatabaseMapping instance to read from a
database instead of the DiffDatabaseMapping provided by the database
manager. There was an issue that Exporter needed to close the connection
to the database which could not be reopened causing the item to fail
executing more than once. Further, the item should export only
uncommitted changes so using a separate DatabaseMapping makes even more
sense.

Re #473