Try new icons for Importer and Exporter
Re #475.

These new icons are inspired from Font Awesome 5's file-export and
file-import, except that there's a database instead of a file.
The goal is to have symmetrical icons for these two items to better
illustrate their goals.

I don't say the icons look amazing but decent. We could try and
improve in the future, maybe with help from a pro.