Run the Importer using a built-in PythonTool specification
Re #476

I think it's important that executing an Importer doesn't block the UI.
I tried QProcess but it wasn't great, so I decided to try using
our built-in mechanism for running tools. Basically now, the
importer item has a dedicated, internal python tool specification
as one of its class attributes. When executed, this tool specification
creates an instance and the instance runs using our QSubProcess class,
exactly as the tool.