Fix unittests
After commit 5b57770d a couple of tests needed some small maintenance.

Maybe more importantly, TestTreeViewFormRemoveMixin was
inheriting from unittest.testcase which is just incorrect.
This is supposed to be a mixin class to include in
TestTreeViewForm. I hope that was the reason we were skipping them.
Now this is fixed