Fix invalid item rank indicator
When project item's execution rank could not be declared Toolbox was
trying to set the rank indicator to 'x'. This failed to work since we
automatically add +1 to the rank so the ranks are 1-based, not 0-based.
We now use None to indicate that the rank is unset and
ProjectItem.set_rank() will draw an 'X'.