Correctly write command line arguments in tool spec editor
Re #482

After eeea7adf the Tool specification stores command line arguments
directly as a list. For this reason, we need to convert them to a
string to show them in the line editor.

The reason for storing the command line arguments as a list
(rather than a space-separated string) was, that I needed to have
keyword arguments with spaces. So with the list approach, I can
effectively isolate my arguments, whereas with the space-separated
string, it is impossible to have arguments with spaces.