Fix bugs in the auto filter of parameter tables
There were plenty of bugs here and there. Now everything is much
easier and consistent.

The auto filter now says which values are accepted rather than which
are rejected, because it's easier to understand this way.
None is used to represent that no values
are accepted by a filter (i.e., all of them are rejected).