Fix bug when using the auto filter in tree view
Re #477

The autofilter menu has a functionality to search by word,
and this wasn't working with parameter values which can be float or int.
So now we just convert them to str before searching.

We also fix related bugs with matching the filtered values and the
actual values from columns.