Add a "cancel import on error" checkbox to importer settings
Sometimes there is need to import data even if the import has errors.
Currently for example it will error when an empty row exists in a data source.

If Cancel on error checkbox is ticked then the import is rolled back.
If it is not checked an error will show that there was errors but the
imported data will still show.