Start using spine engine
The spine_engine package is up at the engine repo. This commit
imports SpineEngine from that package and uses it to run dags.

It entirely replaces ExecutionInstance, ExecutionState,
and ResourceMap. It also means that ProjectItems need to provide
separate execute and resource advertising methods for forward and
upward direction.

There are other changes to match the dagster
architecture, such as waiting for the tool instance to finish in a
local event loop.

Overall I think the code is simpler. Stopping execution needs more
work, and we still need to repair unittests. Probably move some of
them to the spine_engine package.

The spine_engine package has been added to requirements.txt, so I
hope it will be installed by installing toolbox. Haven't tested it.

Re #492