Indexed parameter value expansion for DomainSets
Indexed parameter values need to be expanded to an additional dimension
before they can be exported to a .gdx file. This commit implements the
basic functionality to do so: a DomainSet containing such parameters is
converted (sometimes partially) to a Set with an extra dimension which
is indexed by a given index domain.

Re #490