Try and improve performance of curved links
Re #486

Curved links geometry is determined using sort of a crazy algorithm.
Basically we trace a bezier curve between source and
destination and then give thickness to this curve. The thickness part
is the problem, since there's nothing like that directly from Qt
(or at least I couldn't find it).

So what we do at the moment is traverse the curve finding critical
points where the angle changes more than a hardcoded treshold.
At these points we follow the normal to the curve to find extreme
points of our thick curve.

You can imagine how much computation is this, to perform every time
a few links move in the scene.

So now I'm trying a few tricks, for example, I realized we only need
to compute half of the points since the curves are always symmetrical.