Merge DomainSet and Set
Merged the DomainSet and Set classes in gdx.py into a single Set class.
It does not really make sense to have two separate objects for these as
the only difference is that DomainSets don't have indexing domains. We
just use [None] for the list of domain names in these cases. This
removes a lot of code duplication as we can process both domains and
sets using the same functions now.

Re #490