Use `QProcess.write` to pass importer arguments
Re #494

A bit of history. We want the Importer item to do its thing without
blocking the UI. QThread is not the best option because of this
PySide2 bug: https://bugreports.qt.io/browse/PYSIDE-803, so we should
try and use QProcess.

So we created importer_program.py that we run using QProcess.
Our first approach was to pass arguments in the command line but
we hit some limitations (ARG_MAX in linux, equivalent to 2MB), so
that's also a no go.

Writing a file to disk is awkward, so here I'm trying to use
`QProcess.write` combined with the python builtin `input` at
`importer_program.py` side. It works with our test data so
hopefully this is a win.