Fix bug when overwritten a db with changes
When the user pressed create new spine db on a database being viewed,
we just closed the tree, graph or tabular view. This was a bit
dramatic and actually didn't work well if the db had pending changes.

So now we politely ask the user to first close the open windows and
then press create...