WIP: optimize space in tabular view
Re #507

If we want to merge the tabular view into the tree view we probably
want to reduce the amount of widgets. This commit is a WIP towards
that. The idea is to get rid of the pivots lists and do the pivoting
directly by dragging headers. The headers are now index widgets that
allows one to do both pivoting and filtering.