Fix PivotModel tests
Unfortunately I had to remove a lot of tests since PivotModel was
doing a lot of stuff that now is controlled by other components,
such as add, update and remove data. This made a bunch of tests
obsolete.

I need to come back to this at some point (pretty soon, always
within the context of #461) to make sure we didn't lose test
coverage.