Keep pivot table filter status after insertion/removal
Re #461

When filter menu content changed (due to insertion/removal of
objects or parameters) we were 'reseting' the filter model. This
caused the filter status to be forgotten. Now we have reinstated
the `add_items` and `remove_items` to `FilterCheckboxListModel` so
we can do it properly. Tests have also been updated to reflect
previous changes in the above mentioned class.