Fix broken filter menu for mapping options widget
Solves #531

While working on #461 we deeply modified the FilterMenu class to use
it with the tabular view mixin, totally ignoring that the mapping
options widget was also using it. This commit attempts to fix things
for everybody.

I went with a brute-force approach based on subclassing, which I
believe is most correct since it allows us to develop independent
FilterMenu subclasses as needed.