Improve performance of pivot table parameter edition
Re #461

When editing the pivot table, we know for sure that the
parameter value items we send to the database will be correct.
Thus we do not need to go through the expensive integrity checking
process. This commit implements quick functions to add and
update parameter values known to be correct.