Separate project items and project tree items
To make project items less dependent on the Toolbox GUI, the project
item classes that form the Qt model items for ProjectItemModel were
separated from project items. This also solves the issue that
ProjectItems were not really meant to live in a tree structure. It does
not make much sense to make a data store a child of a tool, for example.
- ProjectItem now inherits directly from MetaObject
- BaseProjectItem, ProjectItemRoot and CategoryProjectItem were renamed
to BaseProjectTreeItem etc. and moved to a separate module,
project_tree_item
- A new project tree item, LeafProjectTreeItem has been introduced which
holds a reference to its ProjectItem object

Re #546