Upgrade Importer to spine project v1
Change absolute file paths that are saved in Importer settings to
dict produced by serialize_path

Change the table name of returned from the CSVConnector to 'csv' instead of
the source path. Add uppgrade script to change the table name to 'csv' in mapping settings
for CSVConnector.

Remove unnecessary eval