Fix project item properties widgets' context menus
Many context menus were broken due to recent refactoring/renaming in
ProjectItemModel where project_item() was renamed to item() and instead
of returning ProjectItem instance it returned LeafProjectTreeItem.

Also, View item's context menu was not up-to-date with the fact that the
properties widget contained only a single button: Open View. This was
fixed as well.

Re #550