Reorganize project items' icon animations
Icon animations are started/stopped from the execute_forward() methods
of project items themselves. This has the issue that it makes the item
execution run GUI code which we would like to avoid for remote
execution.

This commit leverages changes in spine_engine 0.3.0 which allow the
engine to signal the start and stop of an item's execution. The signals
are now used in DesignQGraphicsView which has been made responsible of
the animations. This in turn makes the items' execution free of the
animation code.

Re #541