Retry creating Exporter settings upon activation
After Exporter failed to initialize settings for a database it would
remain in that state until reopening the project even though the
database issue might have been fixed in the meantime. We now retry
reading the initial settings every time the Exporter item is activated
if it is in an erroneous state.

Re #257