Fix issues with path separators
Normalize also absolute paths when loading a project; this should fix
cases where e.g. Data Store URLs have '/' in Windows instead of '\'.

Also make sure that that Exporter uses the correct path separators when
upgrading from old project file format. There was an issue that the Data
Store provided a URL with '\' as separators but this was compared with a
URL using '/' as separator resulting in loosing the original export
settings.

Re #257