Get rid of NotificationStatusBar in favor or DBSessionHistoryDialog
The notification status bar wasn't working too well with the undo/
redo (basically, for each undo/redo action we were adding a new
notification which became quickly unmanageable).

Now we introduce a custom dialog that shows the history of db
changes in the current session, in a QColumnView. Works a bit better
in my opinion.

Update to CHANGELOG comming soon.