Attempt to fix tests on linux
It looks like the maximum length we pass to `elidedText` is in pixels,
which makes the result somewhat platform dependent. As result,
text is elided in the travis environment that is not elided in my
machine for instance. Now I put a hopefully sufficient large
maximum length for the test to pass everywhere.