Minimize dag simulations due to undo/redo
Re #39

When undoing item removal we need to add the items and the
links that were removed with the item. With this commit, we only
simulate the dag after all links have been restablished.

This is mainly so the Exporter don't lose their settings due to
dag simulation, but it's also nicer from the programming point of view.