Implement parameter merging in Exporter
Parameter merging is needed for certain GAMS models. In a nutshell, it
means that a number of parameters from an entity class get merged into a
new parameter and the names of the individual parameters are collected
into a new GAMS domain.

This commit contains the needed changes to the spine_io.exporters.gdx
module and new merging settings GUI for the Exporter project item.

Re #572