Prevent undo/redo to modify the wrong item properties
Re #39

Before this commit, if one changed something in one Tool, selected
another Tool and then undo/redo, the changes were applied on the
currently selected item even though it wasn't concerned at all.

We now check if the concerned item is active before modifying its
properties in the context of undo/redo. This is done by introducing
a new attribute `_active`, which is updated in `ProjectItem.activate`
and `ProjectItem.deactivate` (actually `Exporter` was already doing
this, we now just have it in the base class).