Introduce critical commands
Re #39

Now before closing the project, we check if there any commands in the
undo stack that are marked as critical, and undo them.

This is mainly so we can keep project item and data directory names
properly synchronized. So if the user renames an item and then closes
the project without saving it, we undo the renaming which also switches
the data directory name back to the original.