Fix bug in adding entities to the graph view
When adding entities from the graph itself (Drag and drop from Item
palette, then merging objects) caused the corresponding entities to
be materialized twice in the graph, resulting in ugly bugs.

Now we make sure they're materialized once by using a trick involving
QTimer. It's not so simple, unfortunately, but it gets the job done.