Add Tooltips for items in Tabular view and Treeview
Re #511

This commit adds a tooltiprole path for data methods in single_parameter_models, tabular_view_models
and frozen_table_model. When hoovering a cell that corresponds to an item in the database that has a description
this is now shown in the Qt.ToolTipRole. If no description is availible then display name instead.