Start introducing ParameterViewFilterMenu
Re #596

WIP towards using a `FilterMenuBase` subclass for filtering in the
parameter tables, thus consolidating our filtering interfaces.

At the moment the filter menu does nothing more than
getting populated with the unique items in the corresponding column.
It's possible to lazy load the filter as we lazy load the table,
which frankly is a relief.