Add and remove item from filter menus appropriately
Re #596

It turns out we need to store the 'values' in the filter menu models
rather than the id, since for each 'value' we can have multiple ids
(and even dbs) associated in the parameter models (remember the
parameter models hold data for multiple databases and multiple
entities at the same time, unlike the pivot table model).

But things look better now. The problem is we need to manually
handle item renaming.