WIP towards running SpineDBManager in another thread
The idea is to move SpineDBManager instances to another thread
entirely. With this we can do db-related work without bothering the UI.
This replaces lazy loading: the Data Store form UI shows instantly
while everything is collected in the background.

Looks promising, but now we need to worry about thread affinity.
This commit introduces a few bugs because of this. They will be fixed
in further commits.