Improved heuristics to detect off-screen windows
When restoring window states it is possible that the screen
configuration has changed since last Toolbox launch. This might lead to
situations where a restored window ends up off screen. We do check if
the number of screens is smaller than before and move the window to the
top left corner of the primary screen if it is so. However, this is not
always enough as the off screen issue may materialize when the screen
geometry changes, not the number of screens. We now check if the window
center is within any existing screen after restoring the window to cover
these cases.

Re #615