Introduce SpineDBFetcher
Turns out this is the most sensitive thing to do. Do only the fetching
in a specific class, and all the rest through the signaller. This is
to avoid issues with thread affinity.

Also, we note that querying the database is not the blocking operation,
but populating the Data Store models is.